requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
phonenumberslite==8.12.48
pytz==2022.1
sqlparse==0.4.2
